+++ src/main/java/org/folio/inventorymatch/MainVerticle.java
@@ -29,6 +29,7 @@ public class MainVerticle extends AbstractVerticle {
     Router router = Router.router(vertx);
     router.put("/*").handler(BodyHandler.create()); // Tell vertx we want the whole PUT body in the handler
     router.put(MatchService.INSTANCE_MATCH_PATH).handler(matchService::handleInstanceMatching);
+    router.put(MatchService.INSTANCE_UPSERT_MATCHKEY_PATH).handler(matchService::handleInstanceMatching);
     router.put(MatchService.INSTANCE_UPSERT_HRID_PATH).handler(matchService::handleInstanceUpsertByHrid);
 
     vertx.createHttpServer()